+++ docx/oxml/shape.py
@@ -102,6 +102,68 @@ class CT_Inline(BaseOxmlElement):
         )
 
 
+class CT_Anchor(BaseOxmlElement):
+    """
+    ``<w:anchor>`` element, container for an inline shape.
+    """
+    extent = OneAndOnlyOne('wp:extent')
+    docPr = OneAndOnlyOne('wp:docPr')
+    graphic = OneAndOnlyOne('a:graphic')
+
+    @classmethod
+    def new(cls, cx, cy, shape_id, pic):
+        """
+        Return a new ``<wp:anchor>`` element populated with the values passed
+        as parameters.
+        """
+        anchor = parse_xml(cls._anchor_xml())
+        anchor.extent.cx = cx
+        anchor.extent.cy = cy
+        anchor.docPr.id = shape_id
+        anchor.docPr.name = 'Picture %d' % shape_id
+        anchor.graphic.graphicData.uri = (
+            'http://schemas.openxmlformats.org/drawingml/2006/picture'
+        )
+        anchor.graphic.graphicData._insert_pic(pic)
+        return anchor
+
+    @classmethod
+    def new_pic_anchor(cls, shape_id, rId, filename, cx, cy):
+        """
+        Return a new `wp:anchor` element containing the `pic:pic` element
+        specified by the argument values.
+        """
+        pic_id = 0  # Word doesn't seem to use this, but does not omit it
+        pic = CT_Picture.new(pic_id, filename, rId, cx, cy)
+        anchor = cls.new(cx, cy, shape_id, pic)
+        anchor.graphic.graphicData._insert_pic(pic)
+        return anchor
+
+    @classmethod
+    def _anchor_xml(cls):
+        return (
+            '<wp:anchor distT="0" distB="0" distL="114300" distR="114300" simplePos="0" behindDoc="1" locked="0" layoutInCell="1" allowOverlap="1" %s>\n'
+            '  <wp:simplePos x="0" y="0" />\n'
+            '  <wp:positionH relativeFrom="column" >\n'
+            '    <wp:posOffset> -571500 </wp:posOffset >\n'
+            '  </wp:positionH >\n'
+            '  <wp:positionV relativeFrom="paragraph" >\n'
+            '  <wp:posOffset > -609600 </wp:posOffset >\n'
+            '  </wp:positionV >\n'
+            '  <wp:extent cx="6904934" cy="10033000"/>\n'
+            '  <wp:effectExtent l="0" t="0" r="4445" b="0" />\n'
+            '  <wp:wrapNone />\n'
+            '  <wp:docPr id="666" name="unnamed"/>\n'
+            '  <wp:cNvGraphicFramePr>\n'
+            '    <a:graphicFrameLocks noChangeAspect="1"/>\n'
+            '  </wp:cNvGraphicFramePr>\n'
+            '  <a:graphic>\n'
+            '    <a:graphicData uri="URI not set"/>\n'
+            '  </a:graphic>\n'
+            '</wp:anchor>' % nsdecls('wp', 'a', 'pic', 'r')
+        )
+
+
 class CT_NonVisualDrawingProps(BaseOxmlElement):
     """
     Used for ``<wp:docPr>`` element, and perhaps others. Specifies the id and